+++ code.js
@@ -4,10 +4,48 @@
   var PLUGIN_DATA_NAMESPACE = "codesnippets";
   var PLUGIN_DATA_KEY = "snippets";
 
-  // src/utils.ts
-  function formatString(string, rawString, filter) {
-    if (!filter)
-      filter = "hyphen";
+  // src/snippets.ts
+  var regexSymbols = /\{\{([^\{\?\}\|]+)(\|([^\{\?\}]+))?\}\}/g;
+  var regexQualifierSingle = "([^}&|]+)";
+  var regexQualifierOr = "([^}&]+)";
+  var regexQualifierAnd = "([^}|]+)";
+  var regexQualifiers = [
+    regexQualifierSingle,
+    regexQualifierOr,
+    regexQualifierAnd
+  ].join("|");
+  var regexQualifier = new RegExp(`{{([?!])(${regexQualifiers})}}`, "g");
+  async function snippetDataFromNode(currentNode, paramsMap) {
+    const data = [];
+    const seenTemplates = {};
+    async function pluginDataForNode(node) {
+      const pluginData = node.getSharedPluginData(
+        PLUGIN_DATA_NAMESPACE,
+        PLUGIN_DATA_KEY
+      );
+      if (pluginData && !seenTemplates[pluginData]) {
+        seenTemplates[pluginData] = 1;
+        const { pluginDataArray, codeArray } = await hydrateSnippets(
+          pluginData,
+          paramsMap
+        );
+        data.push({ codeArray, pluginDataArray, nodeType: node.type });
+      }
+    }
+    await pluginDataForNode(currentNode);
+    if (currentNode.type === "INSTANCE") {
+      if (currentNode.mainComponent) {
+        await pluginDataForNode(currentNode.mainComponent);
+        if (currentNode.mainComponent.parent && currentNode.mainComponent.parent.type === "COMPONENT_SET") {
+          await pluginDataForNode(currentNode.mainComponent.parent);
+        }
+      }
+    } else if (currentNode.type === "COMPONENT" && currentNode.parent && currentNode.parent.type === "COMPONENT_SET") {
+      await pluginDataForNode(currentNode.parent);
+    }
+    return data;
+  }
+  function formatStringWithFilter(string, rawString, filter = "hyphen") {
     const splitString2 = string.split("-");
     const capitalize2 = (s) => s.charAt(0).toUpperCase() + s.substring(1);
     switch (filter) {
@@ -26,17 +64,6 @@
     }
     return splitString2.join(" ");
   }
-
-  // src/hydrateSnippets.ts
-  var regexQualifierSingle = "([^}&|]+)";
-  var regexQualifierOr = "([^}&]+)";
-  var regexQualifierAnd = "([^}|]+)";
-  var regexQualifiers = [
-    regexQualifierSingle,
-    regexQualifierOr,
-    regexQualifierAnd
-  ].join("|");
-  var regexQualifier = new RegExp(`{{([?!])(${regexQualifiers})}}`, "g");
   async function hydrateSnippets(pluginData, { raw, params }) {
     const pluginDataArray = JSON.parse(pluginData);
     const codeArray = [];
@@ -48,9 +75,7 @@
         matches.forEach((match) => {
           line = line.replace(match[0], "");
         });
-        const symbolMatches = [
-          ...line.matchAll(/\{\{([^\{\?\}\|]+)(\|([^\{\?\}]+))?\}\}/g)
-        ];
+        const symbolMatches = [...line.matchAll(regexSymbols)];
         if (qualifies && symbolMatches.length) {
           let succeeded = true;
           symbolMatches.forEach((symbolMatch) => {
@@ -58,7 +83,11 @@
               (a) => a ? a.trim() : a
             );
             if (param in params) {
-              const value = formatString(params[param], raw[param], filter);
+              const value = formatStringWithFilter(
+                params[param],
+                raw[param],
+                filter
+              );
               line = line.replace(match, value);
             } else if (param === "figma.children") {
               console.log("HELLO WORLD");
@@ -215,7 +244,7 @@
       params["component.name"] = splitString(componentNode.name);
     }
     for (let key in css) {
-      const k = formatString(key, key, "camel");
+      const k = formatStringWithFilter(key, key, "camel");
       params[`css.${k}`] = css[key];
       raw[`css.${k}`] = css[key];
     }
@@ -336,6 +365,114 @@
     return string.replace(/([a-z])([0-9])/g, "$1 $2").replace(/([-_/])/g, " ").replace(/  +/g, " ").trim().toLowerCase().split(" ").join("-");
   }
 
+  // src/bulk.ts
+  function bulkImport(eventData) {
+    const componentsByKey = getComponentsInFileByKey();
+    const data = JSON.parse(eventData);
+    let componentCount = 0;
+    for (let componentKey in data) {
+      const dataToSave = JSON.stringify(data[componentKey]);
+      const component = componentsByKey[componentKey];
+      if (component) {
+        componentCount++;
+        component.setSharedPluginData(
+          PLUGIN_DATA_NAMESPACE,
+          PLUGIN_DATA_KEY,
+          dataToSave
+        );
+      }
+    }
+    const s = componentCount === 1 ? "" : "s";
+    figma.notify(`Updated ${componentCount} Component${s}`);
+  }
+  function bulkExport() {
+    const jsonString = getExportJSON();
+    figma.ui.postMessage({
+      type: "EXPORT",
+      code: jsonString
+    });
+  }
+  function bulkGetComponentData() {
+    const jsonString = getComponentDataJSON();
+    figma.ui.postMessage({
+      type: "COMPONENT_DATA",
+      code: jsonString
+    });
+  }
+  async function bulkGetNodeData() {
+    const jsonString = await getNodeDataJSON();
+    figma.ui.postMessage({
+      type: "NODE_DATA",
+      code: jsonString
+    });
+  }
+  async function getNodeDataJSON() {
+    const nodes = figma.currentPage.selection;
+    const data = {};
+    await Promise.all(
+      nodes.map(async (node) => {
+        data[keyFromNode(node)] = await paramsFromNode(node);
+        return;
+      })
+    );
+    return JSON.stringify(data, null, 2);
+  }
+  function keyFromNode(node) {
+    return `${node.name} ${node.type} ${node.id}`;
+  }
+  function findComponentNodesInFile() {
+    if (figma.currentPage.parent) {
+      return figma.currentPage.parent.findAllWithCriteria({
+        types: ["COMPONENT", "COMPONENT_SET"]
+      }) || [];
+    }
+    return [];
+  }
+  function getComponentsInFileByKey() {
+    const components = findComponentNodesInFile();
+    const data = {};
+    components.forEach((component) => data[component.key] = component);
+    return data;
+  }
+  function getComponentDataJSON() {
+    const components = findComponentNodesInFile();
+    const componentData = {};
+    const data = components.reduce((into, component) => {
+      if (component.parent && component.parent.type !== "COMPONENT_SET") {
+        const lineage = [];
+        let node = component.parent;
+        if (node) {
+          while (node && node.type !== "PAGE") {
+            lineage.push(node.name);
+            node = node.parent;
+          }
+        }
+        lineage.reverse();
+        into[component.key] = {
+          name: component.name,
+          description: component.description,
+          lineage: lineage.join("/")
+        };
+      }
+      return into;
+    }, componentData);
+    return JSON.stringify(data, null, 2);
+  }
+  function getExportJSON() {
+    const data = {};
+    const components = findComponentNodesInFile();
+    components.forEach((component) => {
+      const pluginData = component.getSharedPluginData(
+        PLUGIN_DATA_NAMESPACE,
+        PLUGIN_DATA_KEY
+      );
+      if (pluginData) {
+        data[component.key] = JSON.parse(pluginData);
+      }
+    });
+    return JSON.stringify(data, null, 2);
+  }
+
   // src/code.ts
   if (figma.mode === "codegen") {
     console.clear();
@@ -360,16 +497,16 @@
     figma.on("selectionchange", () => handleCurrentSelection);
     figma.codegen.on("generate", async () => {
       try {
-        const currentNode = handleCurrentSelection();
-        const paramsMap = await paramsFromNode(currentNode);
         const { detailsMode, defaultSnippet } = figma.codegen.preferences.customSettings;
         const isDetailsMode = detailsMode === "on";
         const hasDefaultMessage = defaultSnippet === "message";
-        const snippetData = await findAndGenerateSelectionSnippetData(
-          currentNode,
-          paramsMap
+        const currentNode = handleCurrentSelection();
+        const paramsMap = await paramsFromNode(currentNode);
+        const snippetData = await snippetDataFromNode(currentNode, paramsMap);
+        const snippets = codegenResultsFromSnippetData(
+          snippetData,
+          isDetailsMode
         );
-        const snippets = snippetsFromSnippetData(snippetData, isDetailsMode);
         if (isDetailsMode) {
           snippets.push({
             title: "Node Params",
@@ -403,41 +540,13 @@
       if (event.type === "INITIALIZE") {
         handleCurrentSelection();
       } else if (event.type === "COMPONENT_DATA") {
-        const jsonString = getComponentDataJSON();
-        figma.ui.postMessage({
-          type: "COMPONENT_DATA",
-          code: jsonString
-        });
+        bulkGetComponentData();
       } else if (event.type === "NODE_DATA") {
-        const jsonString = await getNodeDataJSON();
-        figma.ui.postMessage({
-          type: "NODE_DATA",
-          code: jsonString
-        });
+        await bulkGetNodeData();
       } else if (event.type === "EXPORT") {
-        const jsonString = getExportJSON();
-        figma.ui.postMessage({
-          type: "EXPORT",
-          code: jsonString
-        });
+        bulkExport();
       } else if (event.type === "IMPORT") {
-        const componentsByKey = getComponentsInFileByKey();
-        const data = JSON.parse(event.data);
-        let componentCount = 0;
-        for (let componentKey in data) {
-          const dataToSave = JSON.stringify(data[componentKey]);
-          const component = componentsByKey[componentKey];
-          if (component) {
-            componentCount++;
-            component.setSharedPluginData(
-              PLUGIN_DATA_NAMESPACE,
-              PLUGIN_DATA_KEY,
-              dataToSave
-            );
-          }
-        }
-        const s = componentCount === 1 ? "" : "s";
-        figma.notify(`Updated ${componentCount} Component${s}`);
+        bulkImport(event.data);
       }
     });
     figma.showUI(__uiFiles__.bulk, {
@@ -462,53 +571,23 @@
       themeColors: true
     });
   }
-  function snippetsFromSnippetData(snippetData, isDetailsMode) {
-    const snippets = [];
+  function codegenResultsFromSnippetData(snippetData, isDetailsMode) {
+    const codegenResult = [];
     snippetData.forEach((pluginDataAndParams) => {
       const { codeArray, pluginDataArray, nodeType } = pluginDataAndParams;
       pluginDataArray.forEach(({ title, code: templateCode, language }, i) => {
         const code = codeArray[i];
         if (isDetailsMode) {
-          snippets.push({
+          codegenResult.push({
             title: `${title}: Template (${nodeType})`,
             code: templateCode,
             language: "PLAINTEXT"
           });
         }
-        snippets.push({ title, language, code });
+        codegenResult.push({ title, language, code });
       });
     });
-    return snippets;
-  }
-  async function findAndGenerateSelectionSnippetData(currentNode, paramsMap) {
-    const data = [];
-    const seenTemplates = {};
-    async function pluginDataForNode(node) {
-      const pluginData = node.getSharedPluginData(
-        PLUGIN_DATA_NAMESPACE,
-        PLUGIN_DATA_KEY
-      );
-      if (pluginData && !seenTemplates[pluginData]) {
-        seenTemplates[pluginData] = 1;
-        const { pluginDataArray, codeArray } = await hydrateSnippets(
-          pluginData,
-          paramsMap
-        );
-        data.push({ codeArray, pluginDataArray, nodeType: node.type });
-      }
-    }
-    await pluginDataForNode(currentNode);
-    if (currentNode.type === "INSTANCE") {
-      if (currentNode.mainComponent) {
-        await pluginDataForNode(currentNode.mainComponent);
-        if (currentNode.mainComponent.parent && currentNode.mainComponent.parent.type === "COMPONENT_SET") {
-          await pluginDataForNode(currentNode.mainComponent.parent);
-        }
-      }
-    } else if (currentNode.type === "COMPONENT" && currentNode.parent && currentNode.parent.type === "COMPONENT_SET") {
-      await pluginDataForNode(currentNode.parent);
-    }
-    return data;
+    return codegenResult;
   }
   function handleCurrentSelection() {
     const node = figma.currentPage.selection[0];
@@ -526,71 +605,5 @@
     } catch (e) {
       return node;
     }
-  }
-  function getExportJSON() {
-    const data = {};
-    const components = componentNodesInFile();
-    components.forEach((component) => {
-      const pluginData = component.getSharedPluginData(
-        PLUGIN_DATA_NAMESPACE,
-        PLUGIN_DATA_KEY
-      );
-      if (pluginData) {
-        data[component.key] = JSON.parse(pluginData);
-      }
-    });
-    return JSON.stringify(data, null, 2);
-  }
-  function getComponentDataJSON() {
-    const components = componentNodesInFile();
-    const componentData = {};
-    const data = components.reduce((into, component) => {
-      if (component.parent && component.parent.type !== "COMPONENT_SET") {
-        const lineage = [];
-        let node = component.parent;
-        if (node) {
-          while (node && node.type !== "PAGE") {
-            lineage.push(node.name);
-            node = node.parent;
-          }
-        }
-        lineage.reverse();
-        into[component.key] = {
-          name: component.name,
-          description: component.description,
-          lineage: lineage.join("/")
-        };
-      }
-      return into;
-    }, componentData);
-    return JSON.stringify(data, null, 2);
-  }
-  async function getNodeDataJSON() {
-    const nodes = figma.currentPage.selection;
-    const data = {};
-    await Promise.all(
-      nodes.map(async (node) => {
-        data[keyFromNode(node)] = await paramsFromNode(node);
-        return;
-      })
-    );
-    return JSON.stringify(data, null, 2);
-  }
-  function keyFromNode(node) {
-    return `${node.name} ${node.type} ${node.id}`;
-  }
-  function getComponentsInFileByKey() {
-    const components = componentNodesInFile();
-    const data = {};
-    components.forEach((component) => data[component.key] = component);
-    return data;
-  }
-  function componentNodesInFile() {
-    if (figma.currentPage.parent) {
-      return figma.currentPage.parent.findAllWithCriteria({
-        types: ["COMPONENT", "COMPONENT_SET"]
-      }) || [];
-    }
-    return [];
   }
 })();
