+++ src/snippets.ts
@@ -1,5 +1,6 @@
-import { StringFilter, formatString } from "./utils";
+import { PLUGIN_DATA_KEY, PLUGIN_DATA_NAMESPACE } from "./config";
 
+const regexSymbols = /\{\{([^\{\?\}\|]+)(\|([^\{\?\}]+))?\}\}/g;
 const regexQualifierSingle = "([^}&|]+)";
 const regexQualifierOr = "([^}&]+)";
 const regexQualifierAnd = "([^}|]+)";
@@ -10,7 +11,78 @@ const regexQualifiers = [
 ].join("|");
 const regexQualifier = new RegExp(`\{\{([\?\!])(${regexQualifiers})\}\}`, "g");
 
-export async function hydrateSnippets(
+export async function snippetDataFromNode(
+  currentNode: SceneNode,
+  paramsMap: CodeSnippetParamsMap
+) {
+  const data: SnippetData[] = [];
+  const seenTemplates: { [k: string]: number } = {};
+
+  async function pluginDataForNode(node: SceneNode) {
+    const pluginData = node.getSharedPluginData(
+      PLUGIN_DATA_NAMESPACE,
+      PLUGIN_DATA_KEY
+    );
+    // skipping duplicates. why?
+    // component instances have same pluginData as mainComponent, unless they have override pluginData.
+    if (pluginData && !seenTemplates[pluginData]) {
+      seenTemplates[pluginData] = 1;
+      const { pluginDataArray, codeArray } = await hydrateSnippets(
+        pluginData,
+        paramsMap
+      );
+      data.push({ codeArray, pluginDataArray, nodeType: node.type });
+    }
+  }
+
+  await pluginDataForNode(currentNode);
+  if (currentNode.type === "INSTANCE") {
+    if (currentNode.mainComponent) {
+      await pluginDataForNode(currentNode.mainComponent);
+      if (
+        currentNode.mainComponent.parent &&
+        currentNode.mainComponent.parent.type === "COMPONENT_SET"
+      ) {
+        await pluginDataForNode(currentNode.mainComponent.parent);
+      }
+    }
+  } else if (
+    currentNode.type === "COMPONENT" &&
+    currentNode.parent &&
+    currentNode.parent.type === "COMPONENT_SET"
+  ) {
+    await pluginDataForNode(currentNode.parent);
+  }
+  return data;
+}
+
+export function formatStringWithFilter(
+  string: string,
+  rawString: string,
+  filter: SnippetStringFilter = "hyphen"
+) {
+  const splitString = string.split("-");
+  const capitalize = (s: string) => s.charAt(0).toUpperCase() + s.substring(1);
+  switch (filter) {
+    case "camel":
+      return splitString
+        .map((word, i) => (i === 0 ? word : capitalize(word)))
+        .join("");
+    case "constant":
+      return splitString.join("_").toUpperCase();
+    case "hyphen":
+      return splitString.join("-").toLowerCase();
+    case "pascal":
+      return splitString.map(capitalize).join("");
+    case "raw":
+      return rawString;
+    case "snake":
+      return splitString.join("_").toLowerCase();
+  }
+  return splitString.join(" ");
+}
+
+async function hydrateSnippets(
   pluginData: string,
   { raw, params }: CodeSnippetParamsMap
 ) {
@@ -26,17 +98,19 @@ export async function hydrateSnippets(
         line = line.replace(match[0], "");
       });
 
-      const symbolMatches = [
-        ...line.matchAll(/\{\{([^\{\?\}\|]+)(\|([^\{\?\}]+))?\}\}/g),
-      ];
+      const symbolMatches = [...line.matchAll(regexSymbols)];
       if (qualifies && symbolMatches.length) {
         let succeeded = true;
         symbolMatches.forEach((symbolMatch) => {
           const [match, param, _, filter] = symbolMatch.map((a) =>
             a ? a.trim() : a
-          ) as [string, string, string, StringFilter];
+          ) as [string, string, string, SnippetStringFilter];
           if (param in params) {
-            const value = formatString(params[param], raw[param], filter);
+            const value = formatStringWithFilter(
+              params[param],
+              raw[param],
+              filter
+            );
             line = line.replace(match, value);
           } else if (param === "figma.children") {
             console.log("HELLO WORLD");
